Add Rubocop (#226)
Apply base autocorrections with minimal customization

+++ db/migrate/20221230200725_add_unique_constraint_users_email.rb
@@ -1,14 +1,11 @@
 class AddUniqueConstraintUsersEmail < ActiveRecord::Migration[7.1]
   def change
-
     # Potentially unsafe in production, but ok
     # to add here (only used locally)
 
     # remove former index that does not support
     # unique constraint
-    if index_exists?(:users, :email)
-      remove_index(:users, :email)
-    end
+    remove_index(:users, :email) if index_exists?(:users, :email)
 
     safety_assured do
       add_index :users, [:email], unique: true
